requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
boto3==1.34.84
botocore==1.34.84
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.8
configparser==6.0.1
contextlib2==21.6.0
cryptography==42.0.5
dataclasses-json==0.6.4
Deprecated==1.2.14
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
docx==0.2.4
embedchain==0.1.98
etelemetry==0.3.1
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
fitz==0.0.1.dev2
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
isodate==0.6.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.31
looseversion==1.3.0
lxml==5.1.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.14.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.29.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
schema==0.7.5
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.0
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.5.2
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
tqdm==4.66.2
traits==6.3.2
transformers==4.39.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
